pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "db-dtypes>=1.4.3",
    "google-auth>=2.40.3",
    "google-cloud-bigquery>=3.36.0",
    "pandas>=2.3.2",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.11",
    "streamlit>=1.49.1",
]
